Safari-611.1.10/LayoutTests/media/modern-media-controls/status-label/status-label-text-selection-expected.txt

Testing that text selection is turned off for StatusLabel.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(mediaControls.statusLabel.element).webkitUserSelect is "none"
PASS window.getComputedStyle(mediaControls.statusLabel.element).cursor is "default"

PASS successfullyParsed is true

TEST COMPLETE
